+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added support to ```Input``` library
 - Added support to ```Image Manipulation``` library
 - Added initial support for the ```Migration``` library (advanced migration management coming soon)
+- Added support to ```HTML table``` library
 
 ## [0.0.7]
 ### Added